Project view design change

+++ app/Http/Controllers/ClientContoller.php
@@ -11,12 +11,20 @@ class ClientContoller extends Controller
 {
     public function index()
     {
-        $client = Client::get();
-        $modifiedowners = $client->map(function ($item) {
-            $item->imagePath = asset("/images/client/{$item->image}");
-            return $item;
+        $clients = Client::with(['projects' => function ($query) {
+            $query->select('client_id', \DB::raw('count(*) as total_projects'))->groupBy('client_id');
+        }])->select('id', 'manager_name', 'manager_phone', 'manager_logo', 'owner_name')->get();
+
+        $modifiedClients = $clients->map(function ($client) {
+            $client->managerLogoPath = asset("/images/client/{$client->manager_logo}");
+
+            $client->total_projects = $client->projects->isNotEmpty() ? $client->projects->first()->total_projects : 0;
+
+            unset($client->projects);
+            return $client;
         });
-        return view('client.client', ['clients' => $modifiedowners]);
+
+        return view('client.client', ['clients' => $modifiedClients]);
     }
 
     /**
@@ -30,36 +38,68 @@ class ClientContoller extends Controller
     /**
      * Store a newly created resource in storage.
      */
-
-    public function store(Request $request)
+    public function store(ClientRequest $request)
     {
-        print_r($request->file('manager_logo'));
-        dd($request->file('owner_logo'));
         try {
             $id = $request->input('id');
-            $inputData = $request->input();
+            $inputData = $request->all();
+            $inputData['isSameAsManager'] = $request->has('isSameAsManager') ? 1 : 0;
 
-            if ($request->hasFile('image')) {
-                $image = $request->file('image');
-                $imageName = time() . '.' . $image->getClientOriginalExtension();
-                // Use the move method to save the image
-                $image->move(public_path('images/client'), $imageName);
-                $inputData['image'] = $imageName;
+            if ($request->hasFile('manager_logo')) {
+                $inputData['manager_logo'] = $this->uploadLogo($request->file('manager_logo'));
 
-                // Delete old image if exists
-                $oldImage = Client::where('id', $id)->value('image');
-                if ($oldImage && file_exists(public_path('images/client/' . $oldImage))) {
-                    unlink(public_path('images/client/' . $oldImage));
+                if ($inputData['isSameAsManager']) {
+                    $inputData['owner_logo'] = $inputData['manager_logo'];
                 }
+
+                if (!empty($id)) {
+                    $this->deleteOldImages($id);
+                }
+            }
+
+            if ($request->hasFile('owner_logo') && !$inputData['isSameAsManager']) {
+                $inputData['owner_logo'] = $this->uploadLogo($request->file('owner_logo'));
+            } elseif ($inputData['isSameAsManager'] && $id) {
+                $inputData['owner_logo'] = Client::where('id', $id)->value('manager_logo');
             }
 
             Client::updateOrCreate(['id' => $id], $inputData);
 
             $message = empty($id) ? "Client added successfully" : "Client updated successfully";
 
-            return redirect('clients')->with('message', $message);
+            // return redirect('clients')->with('message', $message);
+            return response()->json(['isStatus' => true, 'message' => $message]);
         } catch (\Throwable $th) {
             return back()->withError($th->getMessage())->withInput();
+        }
+    }
+
+    private function deleteOldImages($id)
+    {
+        $client = Client::findOrFail($id);
+
+        if ($client->manager_logo) {
+            $this->deleteLogo($client->manager_logo);
+        }
+
+        if ($client->owner_logo && $client->owner_logo !== $client->manager_logo) {
+            $this->deleteLogo($client->owner_logo);
+        }
+    }
+
+    private function uploadLogo($file)
+    {
+        $imageName = time() . rand(10, 99) . '.' . $file->getClientOriginalExtension();
+        $file->move(public_path('images/client'), $imageName);
+        return $imageName;
+    }
+
+    private function deleteLogo($imageName)
+    {
+        $path = public_path('images/client/' . $imageName);
+
+        if (file_exists($path)) {
+            unlink($path);
         }
     }
 
@@ -67,7 +107,8 @@ class ClientContoller extends Controller
     {
         try {
             $client = Client::find($id);
-            $client->imagePath = asset("/images/client/{$client->image}");
+            $client->managerLogoPath = asset("/images/client/{$client->manager_logo}");
+            $client->ownerLogoPath = asset("/images/client/{$client->owner_logo}");
             return view('client.clientAdd', ['head_title' => 'Edit', 'button' => 'Update', 'client' => $client]);
         } catch (\Throwable $th) {
             return back()->withError($th->getMessage())->withInput();
@@ -77,19 +118,22 @@ class ClientContoller extends Controller
     public function destroy(string $id)
     {
         try {
+
             $projectClient = Client::findOrFail($id);
 
-            $imagePath = public_path("images/client/{$projectClient->image}");
-
-            if ($imagePath && File::exists($imagePath)) {
-                File::delete($imagePath);
+            // Check if the client has associated projects
+            if ($projectClient->projects()->exists()) {
+                return response()->json(['isStatus' => false, 'message' => 'Cannot delete client. Client has associated projects.']);
             }
 
+            $this->deleteOldImages($id);
             $projectClient->delete();
 
-            return redirect('clients')->with('message', 'Client deleted successfully');
+            // return redirect('clients')->with('message', 'Client deleted successfully');
+            return response()->json(['isStatus' => true, 'message' => 'Client deleted successfully']);
         } catch (\Throwable $th) {
-            return back()->withError($th->getMessage());
+            return response()->json(['isStatus' => false, 'message' => 'Client not deleted successfully']);
+            // return back()->withError($th->getMessage());
         }
     }
 }
